Initial import of CM version of Nexus One Torch.
This is a modified version of Ben Buxton's Nexus One Torch app which
works with the modifications to the CyanogenMod kernel, and doesn't not
require root access to use high brightness.

Renamed to simply "Torch" as it will work on any version of CyanogenMod
which has the kernel modification in place.