Support the Zeppelin camera led.
This device can be managed using /sys/class/leds/flashlight/brightness
It accepts values from 0-255; 0 is off and 255 is max brightness.
Torch apps were setting this to 100 on the Motorola ROM, so that must be normal brightness.

Devices with the zeppelin board include CLIQ XT, Quench, Motomix, and any other MB501.

Change-Id: I58be159197612ade15cd845db8ffd0b073de3164