Support the Zeppelin camera led.
This device can be managed using /sys/class/leds/flashlight/brightness
It accepts values from 0-255; 0 is off and 255 is max brightness.
Torch apps were setting this to 100 on the Motorola ROM, so that must be normal brightness.

Devices with the zeppelin board include CLIQ XT, Quench, Motomix, and any other MB501.

This patch also changes the main DEVICE file to the value previously mentioned, as per
request of Ricardo Cerqueira. Note that this patch was only tested on my MB501, but
it should work fine on other non-motorola phones.

Change-Id: I58be159197612ade15cd845db8ffd0b073de3164