Implement the ability for Torch to use the camera interface to light the camera LED.
This patch is a modification of a KalimAz patch.  I have switched the SurfaceView for a SurfaceTexture this corrects a proble where the power menu widget could not light the torch as the
MainActivity.surfaceView was unavailable. It also simplifies the patch a bit.

SurfaceTexture piece borrowed from http://stackoverflow.com/questions/6414003/using-surfacetexture-in-android

Change-Id: I7780317b52b892b06652634e8ce19f17098162f4