Add Continuous SysFs Write option
On some kernels, there's a "protection feature" that turns the flash
off after a second or so.
To make flash work always, sysfs was overwritten continuously, thus,
not allowing the kernel driver to turn it off.
After commit 090cd7927a190091fb7d4e7c334b617252fa8365, that was broken
(fixing flashlight on expressatt, though), so the best way is to make
this behaviour configurable.
Use "mNeedsContinuousSysFsWrite" in overlay to enable.

Change-Id: I0035c4ba5470ef3b394c784bf5209167a685b002