Torch: apply wakelocks liberally to affected devices
Many devices get bored and fall asleep while you hold your flashlight in
utter fear of what may lurk in the dark. Because ghosts only strike in
the absence of light, we need to protect our users by wakelocking not
only the camera interface to flash, but also sysfs toggles.

Change-Id: I8997e70396eeb3cc16a5f39ca4c5e59b4aafefdb